requirements.txt
aws-cdk-lib==2.23.0
constructs>=10.0.0,<11.0.0
PyYAML==6.0
discord.py==1.7.3
